Fix issue with RFE not returning.
The take_exception logic seems to be incorrectly depending on rfe
(return from exception).  This causes the next PC to get set to 0x0
when trying to return from exception since there is no exception to take
when returning from exception.

Just removing it allows the next PC to pick up the correct next
instruction, instead of the exception vector.