tinyfpga_bx: Fixup synthesis core config
Changes to core:
 - remove fpu and monitor filesets
   * Monitor is not needed as its for simulation.
   * FPU synth is failing due to some flags yosys doesn't support, we can
     investigate later.
 - set top level verilog module to mor1kx

The synthesis runs now, but pnr fails with the below error:

  $ fusesoc run --target=tinyfpga_bx mor1kx
  ERROR: failed to place cell 'traceport_exec_insn_o[3]$sb_io' of type 'SB_IO'
  ERROR: Placing design failed.

I started adding the cells i.e. (traceport_exec_insn_o[3]) to the pcf
file at random pins just to see if we could get PNR to succeed, but I
think we will not have enough pins.

We don't really need the PNR to run so we have some options:
 1. Try to get olof's pnr: none option to work
 2. Keep PNR and add a simple wrapper module like data/toplevel.v to
    wrap mor1kx and wire most things to 1b'0.