dcache: Allow writing during write_pending
Issue #122

This fixes case 1 from issue 122 where writes are missed after
invalidate state.

The change to invalidate_ack to to allow the invalidate_ack to
be asserted after the invalidate happens not before.

For writes we allow forwarding to WRITE state directly from invalidate
at this point pending_write will still be high and we can use that
to initialiate the write.  When we have delays the cpu_req_i may be
low when we get to the write state.